Fix compiler warnings for -Wall -Werror compatibility
This commit addresses three types of compiler warnings:

1. std::ptr_fun deprecation (GCC 13)
   - Replace std::ptr_fun with lambda in psimpl.h
   - Modern C++ lambda is preferred over deprecated adaptor

2. Variable shadowing (MSVC C4457)
   - Rename loop variable 'v' to 'val' in polygon.h
   - Prevents shadowing of function parameter

3. size_t to int conversions (MSVC C4267)
   - Add explicit static_cast<int> for QVector::resize() calls
   - Affects lidar_sensor.cpp, polygon.cpp, triangulator.cpp
   - Qt 5's QVector uses int for size parameters

+++ veranda/Packages/builtin_plugin_logic/pkg-veranda_builtin_sensors/src/lidar_sensor.cpp
@@ -187,7 +187,7 @@ void Lidar_Sensor::_buildModels()
 
     //Build new sensor-hits shapes
     scan_image.clear();
-    scan_image.resize(data->ranges.size());
+    scan_image.resize(static_cast<int>(data->ranges.size()));
 
     double curr_angle = data->angle_min;
     double scan_radius = radius.get().toDouble();
